requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
numpy==1.24.1
boto3
openai-whisper